Edit Patient API - Edit names

+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/file/demographics/name/PatientNameDemographic.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 record PatientNameDemographic(
     @JsonProperty(required = true)
-    long identifier,
+    short identifier,
     @JsonProperty(required = true)
     @JsonDeserialize(using = FormattedLocalDateJsonDeserializer.class)
     LocalDate asOf,
